# feat(performance): Add SSE2 code path to Point