--- a/source/Point.h
+++ b/source/Point.h
@@ -17,6 +17,8 @@ this program. If not, see <https://www.gnu.org/licenses/>.
 
 #ifdef __SSE3__
 #include <pmmintrin.h>
+#elif defined __SSE2__
+#include <xmmintrin.h>
 #endif
 
 
@@ -36,9 +38,8 @@ public:
 	explicit operator bool() const noexcept;
 	bool operator!() const noexcept;
 
-	// No comparison operators are provided because I never expect to use them
-	// and because comparisons with doubles are inherently unsafe due to the
-	// possibility of rounding errors and imprecision.
+	bool operator==(const Point &other) const noexcept;
+	bool operator!=(const Point &other) const noexcept;
 
 	Point operator+(const Point &point) const;
 	Point &operator+=(const Point &point);
@@ -85,7 +86,7 @@ public:
 
 
 private:
-#ifdef __SSE3__
+#ifdef __SSE2__
 	// Private constructor, using a vector.
 	explicit Point(const __m128d &v);
 
@@ -110,7 +111,7 @@ private:
 // Inline accessor functions, for speed:
 inline double &Point::X()
 {
-#ifdef __SSE3__
+#ifdef __SSE2__
 	return val.x;
 #else
 	return x;
@@ -121,7 +122,7 @@ inline double &Point::X()
 
 inline const double &Point::X() const noexcept
 {
-#ifdef __SSE3__
+#ifdef __SSE2__
 	return val.x;
 #else
 	return x;
@@ -132,7 +133,7 @@ inline const double &Point::X() const noexcept
 
 inline double &Point::Y()
 {
-#ifdef __SSE3__
+#ifdef __SSE2__
 	return val.y;
 #else
 	return y;
@@ -143,7 +144,7 @@ inline double &Point::Y()
 
 inline const double &Point::Y() const noexcept
 {
-#ifdef __SSE3__
+#ifdef __SSE2__
 	return val.y;
 #else
 	return y;