**Feature**

This PR speeds up Point operations on all x86_64 CPUs via vectorization.

## Summary
Adds an SSE2 code path to Point in addition to the SSE3 and non-SSE paths. As SSE2 is guaranteed in x86_64, this will speed the game up for the vast majority of users, and will be present in the steam/github builds (unlike SSE3).

New tests were added to cover the entire Point class, both for correctness and for benchmarking.

Some existing SSE3 code paths were improved, as some of them didn't even use any SSE instructions at all.

Benchmark results:
<details>
<summary>Non-SSE</summary>

```
90: -------------------------------------------------------------------------------
90: Benchmark Point::Point()
90: -------------------------------------------------------------------------------
90: /home/user/Documents/GitHub/endless-sky/tests/unit/src/test_point.cpp:407
90: ...............................................................................
90: 
90: benchmark name                       samples       iterations    est run time
90:                                      mean          low mean      high mean
90:                                      std dev       low std dev   high std dev
90: -------------------------------------------------------------------------------
90: Point::Point()                                 100          7560      2.268 ms 
90:                                         3.94918 ns    3.92037 ns    4.02916 ns 
90:                                        0.226328 ns  0.0977104 ns   0.477432 ns 
90:                                                                                
90: Point::Point(x, y)                             100          8830      2.649 ms 
90:                                         3.69467 ns    3.51541 ns    4.36945 ns 
90:                                         1.58834 ns   0.359512 ns    3.70969 ns 
90:                                                                                
90: Point::Point(Point)                            100          8913     2.6739 ms 
90:                                         3.33585 ns    3.30291 ns    3.38451 ns 
90:                                          0.2017 ns   0.148955 ns   0.265722 ns 
90:                                                                                
90: 
90: -------------------------------------------------------------------------------
90: Benchmark Point-boolean conversions
90: -------------------------------------------------------------------------------
90: /home/user/Documents/GitHub/endless-sky/tests/unit/src/test_point.cpp:421
90: ...............................................................................
90: 
90: benchmark name                       samples       iterations    est run time
90:                                      mean          low mean      high mean
90:                                      std dev       low std dev   high std dev
90: -------------------------------------------------------------------------------
90: Zero point                                     100          8835     2.6505 ms 
90:                                         3.27441 ns    3.26016 ns    3.30373 ns 
90:                                        0.099461 ns  0.0571766 ns   0.156599 ns 
90:                                                                                
90: Negate zero point                              100          9569     2.8707 ms 
90:                                         3.07649 ns    3.05412 ns    3.12682 ns 
90:                                        0.162383 ns  0.0895967 ns   0.313365 ns 
90:                                                                                
90: Nonzero point                                  100          8903     2.6709 ms 
90:                                         3.30954 ns    3.28522 ns    3.35302 ns 
90:                                        0.161225 ns   0.103183 ns   0.241221 ns 
90:                                                                                
90: Negate nonzero point                           100          9569     2.8707 ms 
90:                                         3.04835 ns    3.03459 ns    3.07186 ns 
90:                                       0.0897129 ns  0.0590607 ns   0.126919 ns 
90:                                                                                
90: 
90: -------------------------------------------------------------------------------
90: Benchmark Point arithmetics
90: -------------------------------------------------------------------------------
90: /home/user/Documents/GitHub/endless-sky/tests/unit/src/test_point.cpp:438
90: ...............................................................................
90: 
90: benchmark name                       samples       iterations    est run time
90:                                      mean          low mean      high mean
90:                                      std dev       low std dev   high std dev
90: -------------------------------------------------------------------------------
90: Addition                                       100          6903     2.7612 ms 
90:                                         4.24177 ns    4.21218 ns    4.29128 ns 
90:                                        0.190314 ns   0.129408 ns   0.271838 ns 
90:                                                                                
90: Subtraction                                    100          6907     2.7628 ms 
90:                                         4.23815 ns    4.20452 ns    4.33706 ns 
90:                                        0.268262 ns   0.107935 ns   0.581382 ns 
90:                                                                                
90: Lane-wise multiplication                       100          6292     2.5168 ms 
90:                                           4.488 ns    4.45689 ns    4.54449 ns 
90:                                        0.207658 ns   0.130475 ns   0.322389 ns 
90:                                                                                
90: Scalar multiplication                          100          7664     2.2992 ms 
90:                                         3.81419 ns    3.77571 ns     3.8919 ns 
90:                                        0.266473 ns   0.152342 ns     0.4425 ns 
90:                                                                                
90: Scalar division                                100          7819     2.3457 ms 
90:                                         3.79022 ns    3.75851 ns     3.8644 ns 
90:                                        0.233005 ns   0.110997 ns   0.444823 ns 
90:                                                                                
90: 
90: -------------------------------------------------------------------------------
90: Benchmark Point vector arithmetics
90: -------------------------------------------------------------------------------
90: /home/user/Documents/GitHub/endless-sky/tests/unit/src/test_point.cpp:459
90: ...............................................................................
90: 
90: benchmark name                       samples       iterations    est run time
90:                                      mean          low mean      high mean
90:                                      std dev       low std dev   high std dev
90: -------------------------------------------------------------------------------
90: Dot product                                    100          7332     2.9328 ms 
90:                                         4.02721 ns    3.99073 ns     4.1179 ns 
90:                                        0.267855 ns   0.121686 ns   0.538021 ns 
90:                                                                                
90: Cross product                                  100          7362     2.9448 ms 
90:                                         4.00564 ns    3.98039 ns    4.05196 ns 
90:                                        0.169533 ns   0.109925 ns   0.271748 ns 
90:                                                                                
90: Length                                         100          9524     2.8572 ms 
90:                                         3.06325 ns     3.0468 ns    3.09047 ns 
90:                                        0.105561 ns  0.0724249 ns   0.147425 ns 
90:                                                                                
90: LengthSquared                                  100          9619     2.8857 ms 
90:                                         3.03366 ns    3.02061 ns    3.05826 ns 
90:                                       0.0890944 ns  0.0533828 ns   0.131989 ns 
90:                                                                                
90: Unit                                           100          3554     2.8432 ms 
90:                                          4.9363 ns    4.90481 ns    5.00912 ns 
90:                                        0.230241 ns   0.122618 ns   0.438537 ns 
90:                                                                                
90: Distance                                       100          6862     2.7448 ms 
90:                                         4.23571 ns    4.20763 ns    4.30403 ns 
90:                                        0.210699 ns   0.109474 ns   0.400887 ns 
90:                                                                                
90: DistanceSquared                                100          6982     2.7928 ms 
90:                                         4.25094 ns    4.22473 ns    4.29976 ns 
90:                                        0.176961 ns   0.111857 ns   0.271006 ns 
90:                                                                                
90: Lerp                                           100          6259     2.5036 ms 
90:                                         4.71782 ns    4.69319 ns    4.76792 ns 
90:                                        0.171002 ns  0.0995591 ns   0.268426 ns 
90:                                                                                
90: 
90: -------------------------------------------------------------------------------
90: Benchmark Point helpers
90: -------------------------------------------------------------------------------
90: /home/user/Documents/GitHub/endless-sky/tests/unit/src/test_point.cpp:491
90: ...............................................................................
90: 
90: benchmark name                       samples       iterations    est run time
90:                                      mean          low mean      high mean
90:                                      std dev       low std dev   high std dev
90: -------------------------------------------------------------------------------
90: abs                                            100          9570      2.871 ms 
90:                                         3.07326 ns    3.04876 ns    3.12126 ns 
90:                                         0.16916 ns   0.101207 ns   0.260809 ns 
90:                                                                                
90: min                                            100          6856     2.7424 ms 
90:                                         4.25852 ns    4.22303 ns    4.32755 ns 
90:                                        0.242743 ns   0.142281 ns   0.378761 ns 
90:                                                                                
90: max                                            100          6956     2.7824 ms 
90:                                         4.23307 ns    4.21211 ns     4.2801 ns 
90:                                        0.151528 ns  0.0723822 ns   0.264925 ns 
```
</details>
<details>
<summary>Old SSE3</summary>

```
90: -------------------------------------------------------------------------------
90: Benchmark Point::Point()
90: -------------------------------------------------------------------------------
90: /home/user/Documents/GitHub/endless-sky/tests/unit/src/test_point.cpp:407
90: ...............................................................................
90: 
90: benchmark name                       samples       iterations    est run time
90:                                      mean          low mean      high mean
90:                                      std dev       low std dev   high std dev
90: -------------------------------------------------------------------------------
90: Point::Point()                                 100          5775       2.31 ms 
90:                                         4.62927 ns    4.24057 ns    5.36465 ns 
90:                                         2.59484 ns    1.45576 ns    3.99014 ns 
90:                                                                                
90: Point::Point(x, y)                             100          5952     2.3808 ms 
90:                                         4.13205 ns    4.09435 ns     4.2309 ns 
90:                                        0.291956 ns   0.135394 ns   0.610615 ns 
90:                                                                                
90: Point::Point(Point)                            100          6985     2.0955 ms 
90:                                         3.51897 ns    3.50899 ns    3.54283 ns 
90:                                        0.074908 ns  0.0347704 ns   0.134703 ns 
90:                                                                                
90: 
90: -------------------------------------------------------------------------------
90: Benchmark Point-boolean conversions
90: -------------------------------------------------------------------------------
90: /home/user/Documents/GitHub/endless-sky/tests/unit/src/test_point.cpp:421
90: ...............................................................................
90: 
90: benchmark name                       samples       iterations    est run time
90:                                      mean          low mean      high mean
90:                                      std dev       low std dev   high std dev
90: -------------------------------------------------------------------------------
90: Zero point                                     100          6627     1.9881 ms 
90:                                         3.69734 ns    3.67159 ns    3.73898 ns 
90:                                        0.163909 ns   0.112618 ns   0.222811 ns 
90:                                                                                
90: Negate zero point                              100          6818     2.0454 ms 
90:                                          3.5584 ns    3.53658 ns    3.59377 ns 
90:                                        0.139201 ns  0.0971789 ns   0.200856 ns 
90:                                                                                
90: Nonzero point                                  100          6578     1.9734 ms 
90:                                         3.63252 ns    3.61063 ns    3.66985 ns 
90:                                        0.142938 ns  0.0939041 ns   0.208696 ns 
90:                                                                                
90: Negate nonzero point                           100          6685     2.0055 ms 
90:                                         3.67206 ns    3.65087 ns    3.70665 ns 
90:                                        0.134938 ns  0.0920376 ns   0.190375 ns 
90:                                                                                
90: 
90: -------------------------------------------------------------------------------
90: Benchmark Point arithmetics
90: -------------------------------------------------------------------------------
90: /home/user/Documents/GitHub/endless-sky/tests/unit/src/test_point.cpp:438
90: ...............................................................................
90: 
90: benchmark name                       samples       iterations    est run time
90:                                      mean          low mean      high mean
90:                                      std dev       low std dev   high std dev
90: -------------------------------------------------------------------------------
90: Addition                                       100          5788     2.3152 ms 
90:                                         4.23828 ns    4.20116 ns    4.34122 ns 
90:                                         0.28714 ns   0.125365 ns    0.60553 ns 
90:                                                                                
90: Subtraction                                    100          5507     2.2028 ms 
90:                                         4.48449 ns    4.43962 ns    4.59269 ns 
90:                                        0.358789 ns   0.185098 ns   0.582726 ns 
90:                                                                                
90: Lane-wise multiplication                       100          5773     2.3092 ms 
90:                                         4.26275 ns    4.22212 ns    4.34444 ns 
90:                                        0.282945 ns   0.175052 ns   0.507506 ns 
90:                                                                                
90: Scalar multiplication                          100          5816     2.3264 ms 
90:                                         4.25323 ns    4.21886 ns     4.3354 ns 
90:                                        0.255219 ns   0.131936 ns   0.509203 ns 
90:                                                                                
90: Scalar division                                100          5741     2.2964 ms 
90:                                         4.25957 ns    4.21798 ns    4.34779 ns 
90:                                        0.296029 ns   0.146726 ns   0.504453 ns 
90:                                                                                
90: 
90: -------------------------------------------------------------------------------
90: Benchmark Point vector arithmetics
90: -------------------------------------------------------------------------------
90: /home/user/Documents/GitHub/endless-sky/tests/unit/src/test_point.cpp:459
90: ...............................................................................
90: 
90: benchmark name                       samples       iterations    est run time
90:                                      mean          low mean      high mean
90:                                      std dev       low std dev   high std dev
90: -------------------------------------------------------------------------------
90: Dot product                                    100          4911     2.4555 ms 
90:                                         5.06921 ns    5.01756 ns    5.15503 ns 
90:                                        0.331746 ns   0.228323 ns   0.492905 ns 
90:                                                                                
90: Cross product                                  100          5364     2.1456 ms 
90:                                         5.29517 ns    4.85943 ns     6.1066 ns 
90:                                         2.93487 ns    1.75786 ns    4.34803 ns 
90:                                                                                
90: Length                                         100          6166     2.4664 ms 
90:                                         3.98531 ns    3.96883 ns    4.01724 ns 
90:                                         0.11368 ns  0.0662044 ns   0.171767 ns 
90:                                                                                
90: LengthSquared                                  100          6562     1.9686 ms 
…