@TomGoodIdea 2025-04-06
> SSE3 sucks on my CPU? I'm not sure why it's slower even on the shared SSE2/SSE3 code paths, but I could reproduce both results reliably. Maybe there is some extra overhead for SSE3 that makes it slower for the short operations; if it's slower for everyone, we could remove the SSE3 code path.

We should remove it if we don't plan to use it in any official builds.