@tibetiroka 2025-04-06
If steam had the ability to install a different package based on the host capabilities, we could use it. But right now, yeah, it's only for self-built versions.