@TomGoodIdea 2025-04-06
If you want to run ES on a single-core CPU from 2005, I don't expect you to choose the Steam version, for performance reasons.