@TomGoodIdea 2025-04-16
Is this enabled for 32-bit builds? If yes, we should probably add it to the system requirements.