@petervdmeer 2025-04-22
> We should remove it if we don't plan to use it in any official builds.

That is a good point, but I'm also okay with doing that in a separate PR; this PR by itself is already valuable by improving the SSE2 code paths.